Refactor main for multithreading benchmark
Move main functions to dataset_process.cpp and hpp files.
Inclduie them in the Cmakes so that it cvan be used in the
multithread_dataset_processing_benchmark file.
